Introduce Alternate Toolchain gcc and binutils
In the ideal buildroot world, there is a single Toolchain that will be
used to build the entire set of packages used in a given distribution.

In the OpenPower world, however, this rule is not completely true, due
to the need to build for two different target architectures at once
(ppc64le and ppe42). This is accomplished by the p8-pore-toolchain (for
POWER8 systems) and ppe42-toolchain (POWER9) packages, that can also be
provided as part of an external pre-built SDK.

Another use of the same mechanism can be used to define an "Alternative
Toolchain" for those (OpenPower) packages wishing to use it. This is
specially useful if we want to move at different speeds when updating
the compiler/toolchain, e.g.:

 1) Upstream Buildroot supporting newer compiler (which we want to use
    in the SKiroot environment) while some other packages lag behinds in
    supporting it (case in point: hostboot-p8 and occ-p8)

 2) Allowing a low-coupling between Skiroot environment (i.e., built by
    the default buildroot toolchain) and "host images" that we may not
    want to build with a different toolchain (and invalidate tests)

The main idea behind (2) is to allow us to rebase op-build with an
updated buildroot (so that we update Skiroot), while allowing Hostboot,
OCC, HCODE, Skiboot etc to remain exactly the same. This is specially
useful in easying the pain of downstream trees to keep current with the
Skiroot environment (i.e., due to security vulnerabilities in Skiroot
kernel or userland packages) while not "touching" Host image packages
not needing updates.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>