ci: Enable the build and reuse of alternate toolchain
Modify ci/build-all-defconfigs to always build the "Alternate Toolchain"
as part of the SDK, but also re-use it when enabled.

Even if this adds significant time to building the SDK (by re-building
GCC+binutils even for configs that won't require it), the expectation is
that it won't be rebuilt often.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>